backend/requirements.txt
Click>=8.0
Flask>=2.0.0
itsdangerous>=2.0.0
Jinja2>=3.1.2
MarkupSafe>=2.0.0
Werkzeug>=2.0.0
Pillow>=8.0.0
gunicorn>=20.0.4
Flask-Cors>=3.0.10
numpy
joblib
tensorflow==2.16.2
